// projects/netgrif-components-core/src/lib/data-fields/task-ref-field/model/task-ref-dashboard-tile-constants.ts
/**
 * A collection of Task ref field dashboard tile component constants used to reference elements of the underlying processes
 */
export enum TaskRefDashboardTileConstants {
    DASHBOARD_TILE_X = 'x',
    DASHBOARD_TILE_Y = 'y',
    DASHBOARD_TILE_ROWS = 'rows',
    DASHBOARD_TILE_COLS = 'cols',
    DASHBOARD_TILE_PARENT_ID = 'parent_dashboard_id',
    DASHBOARD_TILE_PROCESS_IDENTIFIER = 'dashboard_tile'
}